requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
blessed==1.20.0
cachetools==4.2.4
certifi
charset-normalizer==3.1.0
cirq-core==0.13.1
cirq-google==0.13.1
click==8.1.3
cloudpickle==2.2.1
colorful==0.5.5
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
dm-tree==0.1.8
duet==0.2.7
filelock==3.9.1
flatbuffers==2.0.7
fonttools==4.39.0
frozenlist==1.3.3
gast==0.4.0
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==1.0.0
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
jsonschema==4.17.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
numpy==1.23.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
prometheus-client==0.16.0
protobuf==3.17.3
psutil==5.9.4
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
ray==2.3.0
ray[tune]
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
sympy==1.8
tabulate==0.9.0
tensorboard==2.6.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.15.0
tensorflow-quantum==0.7.2
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
botocore
boto3
